feat(game,player): pass 4 turns before considering that a player will never reconnect
In order to leave more time for a player to reconnect, he/she must be
disconnected during 4 turns before we consider that he/she will never
reconnect. In this case, when computing the remaining players, this
disconnected player is not taken into account any more.

When a player is disconnected, his/her turn is automatically passed.