fix(cache): remove AI name when opening an AI slot
This prevented the game to be created: the player name was still in the
cache so it was still in the frontend. So this opened but empty slot was
considered occupied when processing the CREATE_GAME request from the
frontend. This resulted in a mismatch between the registered and
provided number of players.