fix(api): don't take into account elapsed time lower than 5 seconds
Since between the request arrives (game.active_player.turn_start_time)
and the time we get here some time has passed. Which means, the elapsed
time sent to the frontend could be greater than 0 at the beginning of a
turn. To prevent this, we only consider "big" elapsed times. If the raw
time is lower than the threshold, we discard it.