feat(powers): handle teardown of StealPower
It's meant to return the power to its original state and to cancel
any handicap a passive trump may have applied.
For active trumps, the teardown must take place immediately after the
stolen power is player.
For passive trumps, the teardown must take place when we try to enable
the trump at the beginning of the new turn. This way, the passive trump
is correctly active during the full turn during which it was stolen.