fix(api): prevent attributes to be overriden by another message
When we play, we could get a message, and thus update the message
attribute at any time including while we wait for something else that
will need its original message value later on. To prevent errors, we
don't store the message in the object and derive the request type from
it with a property but we pass it (and the request type) to all
functions.

See https://sentry.io/organizations/jenselme/issues/1682888076/ and
https://sentry.io/organizations/jenselme/issues/1682888065/